Restructure API with routers
