+++ pyproject.toml
@@ -27,4 +27,4 @@ cov = "pytest --cov=src --cov-report=term-missing"
 lint = "uv tool run ruff check ."
 format = "uv tool run ruff format ."
 tc = "uv tool run ty check ."
-apirun = "uv run fastapi dev src/api.py"
+apirun = "uv run fastapi dev src/main.py"
