+++ src/routers/categories.py
@@ -1,10 +1,8 @@
-from typing import Annotated
+from fastapi import APIRouter, HTTPException
+from sqlmodel import select
 
-from decouple import config
-from fastapi import Depends, FastAPI, HTTPException
-from sqlmodel import Session, create_engine, select
-
-from .models import (
+from ..dependencies import SessionDep
+from ..models import (
     Category,
     CategoryCreate,
     CategoryRead,
@@ -12,27 +10,13 @@ from .models import (
     DeleteResponse,
 )
 
-DATABASE_URL = config("DATABASE_URL")
-connect_args = {"check_same_thread": False}
-engine = create_engine(DATABASE_URL, connect_args=connect_args)
+router = APIRouter(
+    prefix="/categories",
+    tags=["categories"],
+)
 
 
-def get_session():
-    with Session(engine) as session:
-        yield session
-
-
-SessionDep = Annotated[Session, Depends(get_session)]
-
-app = FastAPI()
-
-
-@app.get("/")
-def root():
-    return {"message": "Finance Tracker is alive!"}
-
-
-@app.post("/categories/", response_model=CategoryRead)
+@router.post("/", response_model=CategoryRead)
 def create_category(category: CategoryCreate, session: SessionDep):
     db_category = Category.model_validate(category)
     session.add(db_category)
@@ -41,13 +25,13 @@ def create_category(category: CategoryCreate, session: SessionDep):
     return db_category
 
 
-@app.get("/categories/", response_model=list[CategoryRead])
+@router.get("/", response_model=list[CategoryRead])
 def read_categories(session: SessionDep):
     categories = session.exec(select(Category)).all()
     return categories
 
 
-@app.get("/categories/{category_id}", response_model=CategoryRead)
+@router.get("/{category_id}", response_model=CategoryRead)
 def read_category(category_id: int, session: SessionDep):
     category = session.get(Category, category_id)
     if category is None:
@@ -55,7 +39,7 @@ def read_category(category_id: int, session: SessionDep):
     return category
 
 
-@app.patch("/categories/{category_id}", response_model=CategoryRead)
+@router.patch("/{category_id}", response_model=CategoryRead)
 def update_category(category_id: int, category: CategoryUpdate, session: SessionDep):
     db_category = session.get(Category, category_id)
     if db_category is None:
@@ -71,7 +55,7 @@ def update_category(category_id: int, category: CategoryUpdate, session: Session
     return db_category
 
 
-@app.delete("/categories/{category_id}")
+@router.delete("/{category_id}")
 def delete_category(category_id: int, session: SessionDep) -> DeleteResponse:
     db_category = session.get(Category, category_id)
     if db_category is None:
@@ -80,6 +64,3 @@ def delete_category(category_id: int, session: SessionDep) -> DeleteResponse:
     session.delete(db_category)
     session.commit()
     return {"detail": f"Category with ID {category_id} deleted successfully"}
-
-
-# TODO: CRUD operations for Transactions
